Send os to builder compute
With arduino-create-agent 1.1.82 the info command return the
operating system in which the agent is installed.
We then send it to builder-api boards compute to obtain the
urls of the required tools for an upload.
This will allow create-agent in the future to be more simple
and to obtain the urls from builder-api rather than from the parsing
of the (multiple) package_index